app/requirements/base.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
Django==4.0.8
django-cors-headers==3.13.0
django-environ==0.9.0
"drf-yasg[validation]"
Pillow==9.2.0

redis==4.3.4
psycopg2-binary==2.9.4

channels==3.0.5
channels-redis==4.0.0

celery==5.2.7
django-celery-beat==2.3.0